Splitting Problem and Problems component to separate files
